Refactor redux's logic in 2 files
